🚨 fix all newly brought up compiler warnings

+++ src/CircuitOptimizer.cpp
@@ -990,13 +990,13 @@ namespace qc {
             // iterate over qubits in reverse order
             for (auto q = static_cast<std::make_signed_t<Qubit>>(msq); q >= 0; --q) {
                 // nothing to be done for this qubit
-                if (dagIterators.at(q) == dag.at(q).end()) {
+                if (dagIterators.at(static_cast<std::size_t>(q)) == dag.at(static_cast<std::size_t>(q)).end()) {
                     continue;
                 }
                 done = false;
 
                 // get the current operation on the qubit
-                auto& it = dagIterators.at(q);
+                auto& it = dagIterators.at(static_cast<std::size_t>(q));
                 auto& op = **it;
 
                 // warning for classically controlled operations
@@ -1012,16 +1012,16 @@ namespace qc {
                 // check whether the gate can be scheduled, i.e. whether all qubits it acts on are at this operation
                 bool              executable = true;
                 std::vector<bool> actsOn(dag.size());
-                actsOn[q] = true;
+                actsOn[static_cast<std::size_t>(q)] = true;
                 for (std::size_t i = 0; i < dag.size(); ++i) {
                     // actually check in reverse order
                     const auto qb = static_cast<std::make_signed_t<Qubit>>(dag.size() - 1 - i);
                     if (qb != q && op->actsOn(static_cast<Qubit>(qb))) {
-                        actsOn[qb] = true;
+                        actsOn[static_cast<std::size_t>(qb)] = true;
 
-                        assert(dagIterators.at(qb) != dag.at(qb).end());
+                        assert(dagIterators.at(static_cast<std::size_t>(qb)) != dag.at(static_cast<std::size_t>(qb)).end());
                         // check whether operation is executable for the currently considered qubit
-                        if (*dagIterators.at(qb) != *it) {
+                        if (*dagIterators.at(static_cast<std::size_t>(qb)) != *it) {
                             executable = false;
                             break;
                         }
